Updated openstack/openstack Project: openstack/horizon  fa10215b311ca67a999b9805e63fac045a6f21f2
Update and merge the install / deployment docs

These two documents were incredibly outdated, but also had almost
identical concerns. This patch updates the content and merges the two
documents.

Note that this guide is still separate to the quickstart guide, which is
intended for developers. In future patches in this blueprint, the file &
navigation structure will be altered to make this distinction more
clear.

Change-Id: Ic5d111d53ab69a3cd1ab08c4c05b2d676e79258c
Implements: blueprint pike-docs-overhaul