Updated openstack/openstack Project: openstack/api-wg  fa4d5f739dbb0d9f32e97c98a0ddadd9b08e837a
Add guideline about consuming endpoints from catalog

Having gone all the way down the rabbit hole... there is nowhere that
actually describes how to select an endpoint from the catalog. This is
just a description of today and matches what keystoneauth is doing.

There are a couple of places where the process described isn't perfect.
For instance, at the end of endpoint selection we just pick the first
from the list if there are still multiple. While that's not _great_ -
it's keystoneauth's current behavior and changing it to be an error
condition would break many many things. So while an attempt has been
made to make this forward compatible with a completely sane future,
there are a few places where, for historical reasons, we need to keep a
non-optimal behavior. They're noted within.

Change-Id: Ie36e8802eff01a846dd6f0d73a5a656856f5ca7d