Updated openstack/openstack Project: openstack/api-wg  028d6012ab05a132467f316f3574fb6568bc2e57
Describe the publication of service-types-authority data

The service-types-authority data is the canonical description of the
service type names and their aliases, but it needs to be consumable if
it's to be the basis of API consumption processes.

Describe how we publish the data so that consumers can depend on it.

Change-Id: I88a2fe2e20252f91a44bcfedbcff9e5acbe049bc