Updated openstack/openstack Project: openstack/nova  f8c8a9902c8cad7babcaa35a0d2696b870e4fd59
Use CONF.host for powervm nodename

The get_available_nodes method of the powervm virt driver was
returning a string based on the machine type, model, and serial
number of the host. This was inconsistent with other drivers. It also
required a working NovaLink connection, which cannot be guaranteed.
This change replaces the MTMS string with CONF.host.

Change-Id: I82142981281f2c26c98e829c728fa96df7555ffd
Partially-Implements: blueprint powervm-nova-compute-driver