Updated openstack/openstack Project: openstack/horizon  8f84c0726df4698b1b31454fccc52e4d9ca4c655
Enable to set step attribute for number input

angular-schema-form doesn't apply "step" attribute
for number type input.

To increase the number properly in the form by form control,
"step" attribute should be applied to number type input.

e.g.) if set 0.1 to step, form control will be increase the number
as 0.1->0.2->0.3... or decrease it as 1->0.9->0.8... .

This patch fixes it.

Change-Id: Ie837b605f281001ae209e743ea50b25a091dd752
Closes-Bug: #1698086