Updated openstack/openstack Project: openstack/cinder  464e8fc8ea3d02d59cc1b2ccafaabebf92f831f0
api-ref: clarify retype docs around default policy permissions

Cinder's default policy allows the owner of a volume to initiate
a retype, even if they aren't an admin. However, if the volume
is going to be migrated during the retype operation, then Cinder
is going to call the Nova swap volume API, which by default is
admin-only. So if a non-admin user tries to retype and migrate
their volume, which is attached to a server instance, then it's
going to fail with a 403 error from Nova.

En lieu of a more complete solution, like Cinder using an
elevated service token to call Nova, this change just attempts
to document the restriction in the API reference.

Change-Id: I6282a6d319beead979780a33880947987906c2f3
Related-Bug: #1698224