Updated openstack/openstack Project: openstack-infra/system-config  a22e4d3751cf5160d9d95544011f3568e0670b6b
Add mailing list for API Users and SDK authors

As part of the effort of collecting all of the external SDK communities
who have OpenStack support, it became clear that we need a forum to talk
about things. Asking people who are not day-to-day developers inside of
OpenStack to join openstack-dev@ is a bit too much, since the majority
of the people in this group are actually not already working on the
OpenStack project itself. (we're talking about people working on things
like fog, jclouds, gophercloud, openstack4j, php-opencloud and the like)

Similarly, because of a wide variety of community norms in real-time
communication, an IRC channel seems like not the right choice. The goal
here isn't to make them join our dev community- but just a place where we
can be assured of being able to talk amongst each other.

Make a mailing list, which is a reasonable lowest-common-denominator and
a reasonable thing to ask arbitrary humans to add themselves to. Keeping
it specifically scoped should keep it at a traffic level that should not
cause people to shy away from joining it.

Change-Id: Id267f93cbc3e244af37de66cb76835d1852c4780