Updated openstack/openstack Project: openstack/diskimage-builder  250aeb5d219db1a0fc01033168a55f7f61e60c74
Fix mkfs failure when loop device is not ready

There was a race in diskimage-builder where the mkfs call after a
kpartx -avs for the loop device would fail because the device was
not yet ready.  This adds a udevadm settle call after the kpartx
to make sure the udev event queue has cleared.

Change-Id: I90103b59357edebbac7a641e8980cb282d37561b
Closes-Bug: #1698337