Updated openstack/openstack Project: openstack-infra/shade  6b325282d3ab2f848de372a80def36b01b6ba6f4
Retry to fetch paginated volumes if we get 404 for next link

When we get a volume list, it's possible for a volume to disappear
causing the pagination to bork. When that happens we retry to
get the list from scratch for 5 times. If all the attempts fail
we just return what we found.

Change-Id: Ia88d1d8a6b3558f7d5d364a9c78bbf834836d3f7
Signed-off-by: Rosario Di Somma <rosario.disomma@dreamhost.com>