Updated openstack/openstack Project: openstack-infra/project-config  87d2bd2a27a10fa77bbc69ed123df0ddf7abd430
Create the requirements stable maintence team

We've been relying on the stable-maintenance team for a while and while we
had tonyb to help that's worked out, with his absence we've started
falling behind on reviews of the stable branches, even with the
stable-maint team's help.  So we wish to take it on ourselves.  In order
to do so we'd need a new gerrit group to be created, at the moment it'd
be dirk, dims and I on that team.

Change-Id: I3bd814d16625910c94fb41f1414c39d4ca938c22