Updated openstack/openstack Project: openstack-dev/devstack  b9f00f453953e2f2271283c396586a771b88244f
docs: add "kvm on s390x" specific configuration in `local.conf`

The upstream CI runs exclusively on nodes with x86 architectures, but
OpenStack supports even more platforms. One of them is the KVM
on s390x (IBM z systems), which is supported since the *Kilo* release.
This change describes the additional settings in the ``local.conf`` file
to enable Devstack on that platform. This is useful for PoCs.

Change-Id: I943b552ca2e36210ac57f36c16db930eb5e58623