Updated openstack/openstack Project: openstack/tripleo-heat-templates  ab6080f94ddb1a0c4b7441064dd2728c98885ada
Update CloudDomain description

First, this parameter must match what is configured on the
undercloud, so strengthen that language.

There is also now an undercloud.conf parameter that can be used to
set the requisite options on the undercloud services, so just point
users at that rather than trying to explain how to configure the
services manually (which is error-prone and doesn't survive
undercloud updates).

Change-Id: I002cce176e3430473a29e79efde3464bddb24cc7