Updated openstack/openstack Project: openstack/tripleo-heat-templates  96813ba2683552860c68d99931aa77cd6a0bf391
Support config dir for env generator input files

We're not going to want to list every single sample environment in
a single file, so let's also take a directory and just read every
yaml file in it.  This commit adds support for that as well as
some initial environments to demonstrate its use.

Change-Id: If2c608f2a61fc5e16784ab594d23f1fa335e1d3c