Updated openstack/openstack Project: openstack/glance  0325eaf2aca99f284c523a5769a3fc901908151b
Add docs and sample configs for running glance with apache

This commit adds docs and sample configs to glance for running glance
under apache. It also documents the limitations with the 2 common
methods of deploying wsgi scripts with apache, mod_wsgi and
mod_proxy_uwsgi, and why those don't work for the glance use case.

Change-Id: I856e7ada856128def48deda5b063d818f45af336