Updated openstack/openstack Project: openstack/nova  3c667e02c1b68f5c29a5353fa76776f57014181d
Handle uuid in HostAPI.compute_node_get

This allows looking up a compute node by its uuid to uniquely
identify it in a list of cells. A later REST API microversion
change will rely on this.

Note that the existing compute_node_get method is used to
minimize the impact to cells v1 (we would have to plumb through
all new support to cells v1 if we added a new HostAPI method).

Since the HostMapping in the API DB does not have the
ComputeNode.uuid we still have to loop over the cells to
find a match. This could be optimized later if needed.

Part of blueprint service-hyper-uuid-in-api

Change-Id: I705119442552e3d507e8d514218466942c5767f5