Updated openstack/openstack Project: openstack/cinder  db8e25a3e06f4bb5df4bad6003600a3cfb53d438
ScaleIO - Make compatible with next SIO version

There is a new verison of ScaleIO in development and one of the
APIs is changing. While backwards compatibility will exist in the
SIO API for a while, this change supports both the new and existing
calls. Tested with SIO 1.3x, 2.0.x, and the upcoming version

Change-Id: I321b7027e3e177276ceebae72fb556d9700169ef
implements: blueprint scaleio-greater-than-2