Updated openstack/openstack Project: openstack-infra/subunit2sql  ca7b04ed589d21444261ded22bbcf61c95b0e94b
Dailycount should not need build_queue to exist.

Dailycount assumes that there is a metadata item called build_queue
with value gate and if this does not exist in the database, the
tool will fail messily. This patch allows you to specify the md
key and value to be used.

Note, it also treats NAN data by filling with 0 - otherwise
the graph gets very confused.

Patch set 2 -
  Fixed pep8 issues.
  Fixed option ordering issue as per review comments.

Patch set 3 -
  Removed debug print statement.

Change-Id: I51baf2ee15d65e315c8142824fae13d964d9f973
Story: 2001034