Updated openstack/openstack Project: openstack/tempest  7992b2826813929e4f932db0fd72df60e3b813eb
create a non-admin token validation test

The test used to the check token validation was previously
available only as an admin test due to the fact that it
reuqired the creation of a new user. the new test,
"test_validate_token", creates a token based on the default
and ensures that it can be validated.

Change-Id: Ice1a241445d532ee2c4b1ad8d2c4c896d755798d