Updated openstack/openstack Project: openstack-infra/project-config  577ecddeb73776a391b90c4adc646d3f5ff78ba3
Propose a job for keystone rolling upgrades

This commit adds a new keystone jenkins jobs the leverages the work
done in openstack-ansible-os_keystone to perform a rolling upgrade
across containers. For what it's worth, the job also runs a
performance script that measures availability of keystone
authenticate and validate API during the upgrade.

Change-Id: If5ac3f00cff61357b59b3339f20b273415202def