Updated openstack/openstack Project: openstack/nova  0cf2628d40d679c0bab7443a16110f119881cb5f
Provide error message in MismatchError for api-samples tests

When the expected response code during a functional API samples
test is not what you'd expect, because something failed, it's
ridiculously confusing figuring out just what went wrong, especially
when the error from the server isn't in the error dump.

This changes the _verify_response method used during functional
API samples tests to dump the response error message with the
MismatchError in the case that the response code is for an error.

So you can go from something like this:

  testtools.matchers._impl.MismatchError: 200 != 400

To something useful like this:

  testtools.matchers._impl.MismatchError: 200 != 400: \
    {"badRequest": {"message": "Invalid input for query \
                     parameters hypervisor_hostname_pattern. \
                     Value: [u'fake']. [u'fake'] is not of type \
                     'string'",
                    "code": 400}}

Change-Id: I79486f762554b5f05557bc0a80e3f0fa16f83ad1