Updated openstack/openstack Project: openstack/nova  9045a98145f0834446d668e0282ddb1c6f2d136c
Consolidate index and detail methods in HypervisorsController

The index and detail methods are identical except for the detail
flag used when calling _view_hypervisor() and get_links().

This moves that common code into a private method.

No test changes are needed since this is purely a refactor with no
change in behavior.

A later change for blueprint service-hyper-uuid-in-api will be
updating the common code for both index() and detail() requests
so this is an early cleanup patch.

Part of blueprint service-hyper-uuid-in-api

Change-Id: I205e651d1527a265e6a9e815012f55a1e1597765