Updated openstack/openstack Project: openstack/horizon  cbb85753fb89a18fac3b00c834d93e63ab3b8d00
Add charts to show volume quotas on Angular launch instance modal

Added two charts for Number of Volumes and Total Volume Storage quotas
on Angular launch instance modal when cinder is enabled.

The charts reflect the volume usage of the new instances to be created
as the user changes the configuration on the modal.

Updated the chart styling for the charts to align better.

Change-Id: Ie744ada2317624153fcfdf9abdf4d7b26996a35e
Partially-implements: blueprint launch-instance-volume-quotas