Updated openstack/openstack Project: openstack/nova  d59d762fa72e1f92edb4a016fe499d6df6cfcea5
api-ref: fix hypervisor_hostname description for Ironic

The hypervisor_hostname shown in the os-hypervisors and servers
APIs is actually the Ironic node uuid, not a name. This comes
from the Ironic virt driver _node_resource method which eventually
gets used in the resource tracker to set the ComputeNode
hypervisor_hostname field, which is used to also set the instance.node
value in the ResourceTracker later during the instance claim.

The confusion probably lies in the fact that the variable passed
around the code is also referred to as "nodename", but in the case
of Ironic it's not actually a hostname, it's a uuid.

Change-Id: Ic639abe9989b282bad71581bb861314619c0f48b
Closes-Bug: #1698000