Updated openstack/openstack Project: openstack-infra/shade  45e689942676f52f5a405262b41c0d60a0d681ab
Only search for floating ips if the server has them

We have a server dict, which has an addresses dict. If the addresses
dict does not list any floating ips, then there is no need to spend the
API call to look for the fip ids needed to do the delete.

Change-Id: Iaf7583fe29bd197a59ae2728fd52d8de2ac1d411