Updated openstack/openstack Project: openstack/nova  a0f0678f32dff18986c26d1772b6588399dbf09b
remove null_safe_int from module scope

The null_safe_int implementation is not a fortunate one as it returns
empty string when called with None. This can be dangerous as it is not
a int. However there is no generic good value to return so this cannot
be really fixed.

Currently the null_safe_int is only used to convert the progress field
of the Instance object into the legacy instance notifications. This patch
proposes no to break that legacy behavior but also discurage the usage
of this convesion in general by removing null_safe_int from the module
scope and inlining to the the only place it is used.

Change-Id: I58b1c128c27d5ae7922eead95f808ea77b3fcd06