Updated openstack/openstack Project: openstack/ironic  f0e6a07ade8191876fe618fc198fb4d93d0b0e3c
Move _abort_attach_volumes functionality to detach_volumes

This is a followup to the Cinder driver patch [0]. There were several
comments during review that asked why we weren't using
detach_volumes instead of _abort_attach_volumes. This patch is a
proposal to remove the method and merge it's functionality with
the detach_volumes method.

This patch also addresses some nits found in the cinder driver.

[0] b23a176a737f217e9573d6f10c3059d22813b159

Change-Id: Ic5584bc1380a65fdd9a74c174bee63bae1c6a4b3
Parital-bug: #1559691