Updated openstack/openstack Project: openstack-infra/project-config  027026a765c33868389f761b61934fc55b787e78
Barbican: update tests

The gate-barbican-.*-devstack.* jobs are slow and should
only be run on the current stable branches.  With updates
to the current barbican master branch, we can run the faster
gate-barbican-.*-dsvm-functional.* jobs.  Both of these jobs
include the same tests, so we need to run the .*-devstack ones
on the stable branches only, and the .*-dsvm-functional ones
on the current master branch and beyond (i.e. not the current
stable branches).

Adds a nv dsvm-functional postgres test to be run on master
and makes some of the other dsvm-functional jobs voting, to
match the voting behavior of the old devstack jobs.

Removes the kmip-devstack and simple-crypto-devstack-py35 jobs
because they weren't running on the stable branches, and now
will not be running on Pike either.  They are replaced by
kmip-dsvm-functional and simple-crypto-dsvm-functional-py35.

Removes the ubuntu-trusty jobs since Mitaka is EOL.

Co-Authored-By: Nam Nguyen Hoai <namnh@vn.fujitsu.com>
Change-Id: I09d0f0a11b18e277d0491db19dad34d3030f5b93