Updated openstack/openstack Project: openstack-infra/project-config  02d18394b89b43ddb623d4ab828780ec4a370e38
Restore selinux context on unbound forwarding.conf

When we modify the forwarding.conf in our ready script we have to
restore the selinux contexts if on centos and fedora. If we don't the
unbound services fails to read the file (despite normal file system
permissions being fine). Address this by restorecon'ing the file if
restorecon is present (implying this is an selinux enabled host).

Change-Id: Ib8a64b0ba9ee4144a6123d7b8f7c1c8fc9501112