Updated openstack/openstack Project: openstack/nova  c74c6fe7efc32c244d72f03f2f49b6e0b9283954
Provide a hint when _verify_response fails

_verify_response compares the template against both the
API response and the docs API samples, and uses subs provided
by the tests for both. However, the result can be different
when comparing the sample to the template and it can fail in
a very confusing way. The answer may be for the test to override
the generalize_subs() method to provide generic regex subs
for a given key in the template. This updates the error message
to provide that hint to see if you need to use generalize_subs()
in your test.

Change-Id: I906b652f2f251f584d4fc460cd2ad04f9b2fb818